% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{ade4_to_mvar}
\alias{ade4_to_mvar}
\title{Convert an ade4 object into class mVarTable}
\usage{
ade4_to_mvar(ade4_object, tables_to_include)
}
\arguments{
\item{ade4_object}{The output of a call to dudi.* from package \code{ade4}.
This is generally a named list containing tables of multiple projections
of interest. Calling \code{class(ade4_object)} should show
\code{ordination_method dudi}.}

\item{tables_to_include}{A character vector specifying which projections in
the \code{ade4} object to store in the mVarTable object. This vector can
have arbitrary length, the only requirement is that calling
\code{ade4_object[tables_to_include[i]]} should be a data frame.}
}
\value{
mvar_table An object of class mvarTable, storing the tables specified
in \code{tables_to_include} and the eigenvalues from the \code{ade4}
decomposition. The annotation slots of each mVarLayer are the row names of
the projected coordinates.
}
\examples{
 library("ade4")
 data(USArrests)
 arrests_pca <- dudi.pca(USArrests, scannf = FALSE, nf = 2)
 ade4_to_mvar(arrests_pca, c("li", "co"))
 arrests_pco <- dudi.pco(dist(USArrests), scannf = FALSE, nf = 2)
 ade4_to_mvar(arrests_pco, c("li", "co"))

 # Example taken from coinertia() in ade4 package
 data(doubs)
 dudi1 <- dudi.pca(doubs$env, scale = TRUE, scan = FALSE, nf = 3)
 dudi2 <- dudi.pca(doubs$fish, scale = FALSE, scan = FALSE, nf = 2)
 coin1 <- coinertia(dudi1,dudi2, scan = FALSE, nf = 2)
 ade4_to_mvar(coin1, tables_to_include = c("li", "co", "aX", "aY"))
 @export
}
