% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot_count_vector_}
\alias{boot_count_vector_}
\title{Boostrap a count vector (Internal)}
\usage{
boot_count_vector_(x, k = NULL, n = 1)
}
\arguments{
\item{x}{(Required) A vector of counts.}

\item{k}{(Optional) The size of the bootstrap sample. Defaults to sum(x).}

\item{n}{(Optional) Default 1. An integer indicating the number of
boostrap count vectors to return.}
}
\value{
A vector or a matrix of \code{n} vectors boostrapped from \code{x}.
}
\description{
Boostrap a count vector (Internal)
}
\examples{
x <- sample(1:1000)
boot_count_vector_(x, n = 5)
boot_count_vector_(x, k = 100, n = 5)
}
