% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot_ordination}
\alias{boot_ordination}
\title{Boostrap ordination.}
\usage{
boot_ordination(D, n = 50, method = "ade4_pca", dist_method = "euclidean",
  rows_annot = NULL, cols_annot = NULL, table_names = NULL,
  common_depth = FALSE, replace_zero = FALSE, round = FALSE, ...)
}
\arguments{
\item{D}{(Required). A data frame of raw counts/weights. Raw data 
is required, and distance matrix/objects are not acceptable.}

\item{n}{(Optional). Default 50. An integer indicating the number of
boostrap samples generated.}

\item{method}{(Required). The method among those listed above that will 
perform the required ordination.}

\item{dist_method}{(Optional). If a distance matrix is used by the specified 
method. We will call \code{vegdist} on the \code{D} using this string as 
the distance.}
}
\value{
A list ordination objects generated by calling 
\code{mvarVis::ordi} on each of the boostrap samples, and then
rotating the coordinates using procustes method to fit them to the 
coordinates of ordination of the original data table.
}
\description{
\code{boot_ordination} computes the ordination for samples boostrapped
from data in \code{D}. Ordination is performed using \code{mvarVis::ordi}
function with specified \code{method}.
}
\examples{
D <-  matrix(runif(100, max = 100), nrow = 25)
bootOrd <- boot_ordination(D, n = 50, method = "ade4_pca", 
                           dist_method = "euclidean")
                           
}

