% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot_prop_vector}
\alias{boot_prop_vector}
\title{Boostrap a vector of proportions}
\usage{
boot_prop_vector(x, n = 1, depth = NULL, replace_zeros = FALSE,
  replace_value = 1)
}
\arguments{
\item{x}{(Required). A vector of counts.}

\item{n}{(Optional). Default 1. An integer indicating the number of
boostrap count vectors to return.}

\item{replace_value}{(Optional) The value to replace zeros with, when
replace_zero is TRUE; i.e. we add a small positive weight. Default value is
1.}

\item{replace_zero}{(Optional) A logical specifying whether to replace
zeros in x.}
}
\value{
A vector or a matrix of \code{n} vectors boostrapped from \code{x}.
}
\description{
\code{boot_prop_vector} returns \code{n} samples of boostraped vector of
proportions. It draws a Dirichlet sample with parameter \code{x}. If
\code{depth} is supplied, the samples are scaled to have sum equal to
\code{depth}.
}
\examples{
y <- runif(10)
y <- y / sum(y)
boot_prop_vector(y, n = 4, replace_zero = 0.05)
}
