% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot_vector}
\alias{boot_vector}
\title{Boostrap a vector}
\usage{
boot_vector(x, n = 1, depth = FALSE, replace_zero = FALSE)
}
\arguments{
\item{x}{(Required). A vector of counts.}

\item{n}{(Optional). Default 1. An integer indicating the number of
boostrap count vectors to return.}

\item{replace_zero}{(Optional). A logical or numeric scalar 
(FALSE by default). If TRUE or numeric zeros in \code{x} will be replaced 
by 1 or \code{replace_zero} respectively, i.e. we add a small positive weight
in place of zeros.}
}
\value{
A vector or a matrix of \code{n} vectors boostrapped from \code{x}.
}
\description{
\code{boot_vector} returns \code{n} samples of boostraped vector of 
raw counts or proportions. If raw counts (integers) are supplied in \code{x},
a multinomial sample is drawn with \code{sum(x)} trials with weights 
equal \code{x}. If \code{x} does not contain integers, a Dirichlet sample
with parameter \code{x}. If \code{depth} is supplied, the samples 
are scaled to have sum equal to \code{depth}.
}
\examples{
boot_vector(sample(1:1000, 5))

x <- sample(1:1000, 10)
x[sample(1:length(x), 4)] <- 0
boot_vector(x, n = 3, replace_zero = TRUE)

y <- runif(10)
y <- y/sum(y)
boot_vector(y, n = 4, replace_zero = 0.05)
}

