% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{build_layers_list}
\alias{build_layers_list}
\title{Construct Layers List}
\usage{
build_layers_list(n_tables, layers_list = "point")
}
\arguments{
\item{n_tables}{How many tables are in the mvarTable object to plot?}

\item{layers_list}{A string specifying the type of layers to include.}
}
\value{
layers_list A list of lists, whose i^th element is a list describing
 the ggplot layers to include for the i^th table.
}
\description{
Here are some defaults for layers options that are commonly
used. The pre-set options currently are
   points: We plot points for each layer. If no color aesthetic associated
     with row or column annotation is provided, each layer of points will
     be a different color. This is the default plot.
   text: This plots text for every layer.
   point-text: This plots points for the first layer and text for the second
     layer. Only applies to tables with two layers.
   text-point: This plots text for the first layer and points for the second.
   points-and-text: This shows both points and slightly offset points for
     every layer.
   point-text-arrow: This plots points for the first layer, and both text
     and arrows for the second layer.
}

