% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/opts.R
\name{build_opts}
\alias{build_opts}
\title{Build Opts lists}
\usage{
build_opts(n_tables, layers_list, aes_list, non_aes_list, facet_vector = NULL)
}
\arguments{
\item{n_tables}{The number of tables for which to construct opts.}

\item{layers_list}{A list containing which layer attributes to plot for
every layer. See the function \code{build_layers_list()} for more
details.}

\item{aes_list}{A list containing the aesthetics attributes for every
layer. See \code{build_aes_and_non_aes_lits()} for more details.}

\item{non_aes_list}{The list containing the non-data-driven aesthetics
attributes for every layer. See \code{build_aes_and_non_aes_lits()} for
more details.}

\item{facet_vector}{A vector containing column names in the annotation data
to use for faceting.}
}
\value{
opts A list that can be input into \code{plot_mvar_from_opts()}.
}
\description{
Build Opts lists
}

