% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvarClass.R
\name{check_table}
\alias{check_table}
\title{Check if valid table format}
\usage{
check_table(object)
}
\description{
A function checking if the object has a valid format
for the table slot of mvarTable.
}

