% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{convert_to_mvar}
\alias{convert_to_mvar}
\title{Convert vegan and ade4 objects to class mvar}
\usage{
convert_to_mvar(X_ord, table_names = NULL)
}
\arguments{
\item{X_ord}{The result of a call to a \code{ade4} or \code{vegan} ordination
method.}

\item{table_names}{A vector of strings specifying which tables to extract, for
\code{ade4} objects. Each of these tables will be an element in the resulting
mvar object. Defaults to c("li", "co").}
}
\value{
An mvar object with the scores and eigenvalues of \code{X_ord}.
}
\description{
Convert an \code{ade4} or \code{vegan} object to class
\code{mvar}.
}

