% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvar_utils.R
\name{cosines}
\alias{cosines}
\title{Get squared cosines from a set of scores}
\usage{
cosines(coord)
}
\arguments{
\item{coord}{The coordinates from which to compute contributions. We assume
coordinates for all dimensions are present.}
}
\value{
The squared cosines between each observation and each component.
}
\description{
Get squared cosines from a set of scores
}
\references{
Abdi and Williams "Principle Component Analysis", equation (11).
}

