% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{default_table_names}
\alias{default_table_names}
\title{Get the tables to put into an mvarTable object}
\usage{
default_table_names(method)
}
\arguments{
\item{method}{The name of the ordination method from which we want to
extract scores.}
}
\value{
If the method is implemented in ade4, the names of the tables that
 give the row and column scores.
}
\description{
For ade4, we need to specify the row and columns scores to
output, depending on the method.
}

