% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/conversion.R
\name{factominer_to_mvar}
\alias{factominer_to_mvar}
\title{Convert a FactoMineR object to class mvarTable}
\usage{
factominer_to_mvar(factominer_object, tables_to_include)
}
\arguments{
\item{tables_to_include}{A character vector specifying which elements in
the \code{FactoMineR} object to store in the mVarTable. We expect that
ade4_object[[cur_table]]$coord be nonnull, for every cur_table in
tables_to_include.}

\item{ade4_object}{The output of a call to a function in FactoMineR.}
}
\description{
Convert a FactoMineR object to class mvarTable
}
\examples{
library("FactoMineR")
example(CA)
factominer_to_mvar(res.ca, c("row", "col"))
example(MFA)
factominer_to_mvar(res, c("ind", "quanti.var"))
}

