% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_annotation_list.R
\name{get_annotation_list}
\alias{get_annotation_list}
\title{Get Annotation to combine with mvarVis Object}
\usage{
get_annotation_list(X_mvar, rows_annot = NULL, cols_annot = NULL)
}
\arguments{
\item{X_mvar}{An object of class mvar, whose annotation elements we want to
fill out.}

\item{rows_annot}{For one table methods, a data.frame whose j^th row
describes the j^th sample in the coordinates matrices corresponding to
row scores. For procustes and CCorA, we require a list of dataframes, one
for each of the directions.}

\item{cols_annot}{The analogous object for column scores.}
}
\value{
A list copying the different annotation data frames so that if the
i^th table in \code{X_mvar} corresponds to row scores, the i^th element in
the list gives the rows annotation, and analogously for column annotation.
}
\description{
Get Annotation to combine with mvarVis Object
}

