% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_ordi_method.R
\name{match_ordi_method}
\alias{match_ordi_method}
\title{Search Vegan and ade4 for matching method}
\usage{
match_ordi_method(method)
}
\arguments{
\item{A}{string describing the function to return. The currently accepted
strings, along with the functions they return, are
  pca: dudi.pca
  pco: dudi.pco
  acm: dudi.acm
  coa: dudi.coa
  fpca: dudi.fpca
  hillsmith: dudi.hillsmith
  mix: dudi.mix
  nsc: dudi.nsc
  vegan_cca: cca, from vegan
and any other string that exactly matches a function in ade4 or vegan.}
}
\value{
ordi_method The function from either ade4 or vegan giving the
 desired function.
}
\description{
Given a character string describing an ordination method
 from ade4 or vegan, return the actual function for that method.
}
