% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mvar_d3.R
\name{merge_mvar_d3_defaults}
\alias{merge_mvar_d3_defaults}
\title{Merge default opts for mvar htmlwidgets plots}
\usage{
merge_mvar_d3_defaults(opts = list(), mvar_object)
}
\arguments{
\item{opts}{A potentially partially specified list, to fill in with defaults.
Currently supported options are, \cr
  $height How tall should each individual table plot be? \cr
  $width How wide should each individual table plot be? \cr
  $asp If one of height or width is ommitted, what is the aspect ratio to
  use to fill that in? \cr
  $types What types of figure should be shown, for each table? Defaults to
  points for all tables.}
}
\value{
A version of opts with defaults filled in.
}

