% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/opts.R
\name{merge_table_plot_opts}
\alias{merge_table_plot_opts}
\title{Supply Default Table Plotting Options}
\usage{
merge_table_plot_opts(opts = list())
}
\arguments{
\item{opts}{A list of plotting options, for which we need to merge
unspecified elements.}
}
\value{
opts An updated list of plotting options.
}
\description{
Supply Default Table Plotting Options
}

