% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvar_utils.R
\name{mvar_annotate}
\alias{mvar_annotate}
\title{Add annotation to an \code{mvarTable}}
\usage{
mvar_annotate(mvar_object, annotations_list)
}
\arguments{
\item{mvar_object}{The \code{mvarTable} whose annotation slots we want to
expand.}

\item{annotations_list}{A list of data frames that will be \code{cbind}ed
into the annotation slots in the \code{mvarTable}.}
}
\value{
mvar_object The original mvar_object, with annotation extended using
   the annotations_list data frames.
}
\description{
Given an \code{mvarTable}, we may want to enrich the annotation
   using separate data frames. This function takes a list of data frames of
   the same length as the number of tables in the \code{mvarTable}, and
   \code{cbind}'s these data frames with the existing annotation slots.
}
