% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvarClass.R
\name{mvar_boot2Table}
\alias{mvar_boot2Table}
\title{Convert mvarBootTable to mvarTable}
\usage{
mvar_boot2Table(mvarBootTab)
}
\description{
Converts mvarBootTable object into 2 mvarTable objects
one for center and one for bootstrap samples.
}

