% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvarClass.R
\name{mvar_boot_to_table}
\alias{mvar_boot_to_table}
\title{Convert mvarBootTable to mvarTable}
\usage{
mvar_boot_to_table(mvar_boot_tab)
}
\description{
Converts mvarBootTable object into 2 mvarTable objects
one for center and one for bootstrap samples.
}
