% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvarClass.R
\docType{class}
\name{mvarBootTable-class}
\alias{mvarBootTable-class}
\title{mvarBootTable-class}
\description{
A class for storing multiple projections of the data set and
of its boostrap samples.
}

