% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ordinate.R
\name{ordi}
\alias{ordi}
\title{Wrapper for ade4 and vegan ordi functions}
\usage{
ordi(X, rows_annot = NULL, cols_annot = NULL, method = "pca",
  dist_method = "euclidean", table_names = NULL, ...)
}
\arguments{
\item{X}{Either a data frame, a distance object, a list of both a data frame
and a distance, a formula and a data frame, or a list of data frames,
depending on the desired method. \code{pcm}, \code{acm}, \code{coa},
\code{fca}, \code{fpca}, \code{hillsmith}, \code{mix},
and \code{nsc} require the raw data frame. \code{pco}, \code{isomap}, and
\code{metaMDS} can take either a \code{data.frame} or a \code{dist}.
vegan_cca and rda require a list whose first element
is a formula and whose second element is a data set containing the
constraining variables, as in the usual \code{cca} function in \code{vegan}.
\code{dpcoa} requires a list whose first element is a data frame and whose
second element is a distance between the rows on that data frame.
\code{CCorA}, \code{ade4_cca}, and \code{procuste} all require a list of
data frames.}

\item{rows_annot}{For one table methods, a data.frame whose j^th row
describes the j^th sample in the coordinates matrices corresponding to
row scores. For procustes and CCorA, we require a list of dataframes, one
for each of the directions.}

\item{cols_annot}{The analogous object for column scores.}

\item{method}{The method among those listed above that will perform the
required ordination.}

\item{dist_method}{If a distance matrix is used by the specified method and
\code{X} is not a distance object, we will call \code{vegdist} on the
\code{X} using this string as the distance.#'}

\item{table_names}{A vector of strings specifying which tables to extract, for}

\item{rows_annot}{}
}
\description{
Perform an ordination using an ade4 or vegan function, and
 return an mvarVis object with the annotation completed.
}

