% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ordi_wrapper.R
\name{ordi_wrapper}
\alias{ordi_wrapper}
\title{Wrapper for vegan and ade4 ordi methods}
\usage{
ordi_wrapper(X, method = "pca", dist_method = "euclidean", ...)
}
\arguments{
\item{X}{Either a data frame, a distance object, a list of both a data frame
and a distance, a formula and a data frame, or a list of data frames,
depending on the desired method. \code{pcm}, \code{acm}, \code{coa},
\code{fca}, \code{fpca}, \code{hillsmith}, \code{mix},
and \code{nsc} require the raw data frame. \code{pco}, \code{isomap}, and
\code{metaMDS} can take either a \code{data.frame} or a \code{dist}.
vegan_cca and rda require a list whose first element
is a formula and whose second element is a data set containing the
constraining variables, as in the usual \code{cca} function in \code{vegan}.
\code{dpcoa} requires a list whose first element is a data frame and whose
second element is a distance between the rows on that data frame.
\code{CCorA}, \code{ade4_cca}, and \code{procuste} all require a list of
data frames.}

\item{method}{The method among those listed above that will perform the
required ordination.}

\item{dist_method}{If a distance matrix is used by the specified method and
 \code{X} is not a distance object, we will call \code{vegdist} on the
 \code{X} using this string as the distance.}
}
\value{
The result of a call to the specified method, in the original
\code{ade4} or \code{vegan} classes.
}
\description{
Given data X and a string describing a method in \code{ade4} or
 \code{vegan}, return the corresponding ordination object from those packages.
}

