% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_mvar_from_opts.R
\name{plot_mvar_from_opts}
\alias{plot_mvar_from_opts}
\title{Plot an mvar object from options list}
\usage{
plot_mvar_from_opts(mvar_object, opts = NULL, opts_center = NULL)
}
\arguments{
\item{mvar_object}{An object of class \code{mvarLayer} that we want to
visualize.}

\item{opts}{A list whose i^th component contains the opts argument for the
i^th table in the \code{mvar_object}. See \code{plot_table()} for a
description of what these need to contain.}
}
\value{
p A ggplot object mapping the layers specified in the arguments.
}
\description{
\code{plot_mvar} helps plot multiple projection layers
associated with an mvar object, using \code{ggplot2} with various options to
customize the plot's appearance.
}

