% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Sample from Dirichlet Distribution}
\usage{
rdirichlet(n = 1, alpha = 1)
}
\arguments{
\item{n}{The number of Dirichlet samples to draw.}

\item{alpha}{(Optional) A vector of parameters for Dirichlet variable.
Defaults to 1.}
}
\value{
A vector of a random sample from Dirichlet distribution with
parameter \code{alpha}
}
\description{
Generate a sample from a dirichlet distribution with parameter
\code{alpha}
}
\examples{
rdirichlet(5, sample(1:10))
}

