% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvar_utils.R
\name{running_cosines}
\alias{running_cosines}
\title{Get running sum of squared cosines from a set of scores}
\usage{
running_cosines(coord)
}
\arguments{
\item{coord}{The coordinates from which to compute contributions. We assume
coordinates for all dimensions are present.}
}
\value{
The sum of squared cosines up to each dimension, for each
observation.
}
\examples{
data(wine)
wine_mfa <- ordi(wine[, -c(1:2)], "MFA", wine[, 1:2], group = c(5,3,10,9,2), type = rep("s",5), graph = F)
coord <- slot(slot(wine_mfa, "table")$ind, "coord")
running_cosines(coord)
}
\references{
Abdi and Williams "Principle Component Analysis", equation (11).
}
