% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvar_utils.R
\name{scale_to_max_radius}
\alias{scale_to_max_radius}
\title{Scale one table to the max radius of another}
\usage{
scale_to_max_radius(x1, x2, q = 0.9)
}
\arguments{
\item{x1}{The reference matrix used to rescale \code{x2}.}

\item{x2}{The matrix to rescale.}

\item{q}{We rescale to this fraction of the maximum radius in x1.}
}
\value{
x2_scaled The rescaled matrix \code{x2}.
}

