% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mvar_utils.R
\name{scale_to_median_radius}
\alias{scale_to_median_radius}
\title{Scale one table to the median radius of another}
\usage{
scale_to_median_radius(x1, x2)
}
\arguments{
\item{x1}{The reference matrix used to rescale \code{x2}.}

\item{x2}{The matrix to rescale.}
}
\value{
x2_scaled The matrix \code{x2} rescaled to have the same median
   radius as \code{x1}.
}
\description{
When making a biplot, the scales of the two sets of points
   aren't always comparable -- only the *directions* of the projections of
   feature variables are meaningful. This helper rescales the points in the
   matrix \code{x2} to have radius comparable to the points in x1.
   Specifically, we rescale \code{x2} so it's median radius is equal to the
   median radius of the \code{x1} points.
}

