% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{vegan_to_mvar}
\alias{vegan_to_mvar}
\title{Convert a vegan object into class mVarTable}
\usage{
vegan_to_mvar(vegan_object, tables_to_include)
}
\arguments{
\item{tables_to_include}{A character vector specifying which elements in
the \code{vegan} object to store in the mVarTable. We expect that
vegan_object[[cur_table]]$coord be nonnull, for every cur_table in
tables_to_include.}

\item{An}{object resulting from a call to a \code{vegan} ordination method.}
}
\value{
mvar_table An object of class \code{mvarTable}, storing the site
   and species tables in the \code{table} slot and the eigenvalues in the
   \code{eig} slot. The annotation slots of each mVarLayer are the row names
   of the projected coordinates.
}
