% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot_ordination}
\alias{boot_ordination}
\title{Boostrap ordination.}
\usage{
boot_ordination(D, n = 50, method = "ade4_pca", dist_method = "euclidean",
  rows_annot = NULL, cols_annot = NULL, table_names = NULL,
  common_depth = FALSE, common_value = NULL, replace_zero = FALSE,
  replace_value = 1, round_values = FALSE, ...)
}
\arguments{
\item{D}{(Required). A data frame of raw counts/weights. Raw data
is required, and distance matrix/objects are not acceptable.}

\item{n}{(Optional). Default 50. An integer indicating the number of
boostrap samples generated.}

\item{method}{(Required). The method that will perform the required
ordination. See ordi()for options.}

\item{dist_method}{(Optional). If a distance matrix is used by the specified
method. We will call \code{vegdist} on the \code{D} using this string as
the distance. Defaults to euclidean distance.}

\item{common_depth}{(Optional). Default \code{FALSE}. The value to which to
normalize the boostrapped column sums. If not provided, do not normalize to
a common depth.}

\item{common_value}{(Optional) The depth to use when common_depth is TRUE.
Defaults to the median of the column sums in tab.}

\item{replace_zero}{(Optional) A logical specifying whether to replace
zeros in x.}

\item{replace_value}{(Optional) The value to replace zeros with, when
replace_zero is TRUE; i.e. we add a small positive weight. Default value is
1.}

\item{round_values}{(Optional). Default FALSE. A logical scalar. Should the
boostrap counts be rounded to the nearest integer?}
}
\value{
A list ordination objects generated by calling
\code{mvarVis::ordi} on each of the boostrap samples, and then
rotating the coordinates using procustes method to fit them to the
coordinates of ordination of the original data table.
}
\description{
\code{boot_ordination} computes the ordination for samples boostrapped
from data in \code{D}. Ordination is performed using \code{mvarVis::ordi}
function with specified \code{method}.
}
\examples{
D <-  matrix(runif(100, max = 100), nrow = 25)
bootOrd <- boot_ordination(D, n = 50, method = "ade4_pca",
                           dist_method = "euclidean", scannf = F, nf = 2)

}

