% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot_table}
\alias{boot_table}
\title{Bootstrap table.}
\usage{
boot_table(tab, n = 1, common_depth = FALSE, common_value = NULL,
  replace_zero = FALSE, replace_value = 1, round_values = FALSE)
}
\arguments{
\item{tab}{(Required). A matrix or data.frame of numeric counts/weights.}

\item{n}{(Optional). Default 1. An integer indicating the number of
boostrap count tables to return.}

\item{common_depth}{(Optional). Default \code{FALSE}. The value to which to
normalize the boostrapped column sums. If not provided, do not normalize to
a common depth.}

\item{common_value}{(Optional) The depth to use when common_depth is TRUE.
Defaults to the median of the column sums in tab.}

\item{replace_zero}{(Optional) A logical specifying whether to replace
zeros in x.}

\item{replace_value}{(Optional) The value to replace zeros with, when
replace_zero is TRUE; i.e. we add a small positive weight. Default value is
1.}

\item{round_values}{(Optional). Default FALSE. A logical scalar. Should the
boostrap counts be rounded to the nearest integer?}
}
\value{
\code{n} x dim(\code{tab})[1] x dim(\code{tab})[2] 3D array of
samples boostrapped from \code{tab}.
}
\description{
\code{boot_table} returns samples boostrapped from table \code{tab}
supplied. Each column of the returned table is a boostrap trial of
the corresponding column of the \code{tab} input table.
The columns of the output table are multinomial samples with weights
proportional to corresponding columns of \code{tab}.
}
\examples{
x <- matrix(sample(1:1000, 30, replace = TRUE), 5, 6)
x[1, 4] <- 0; x[2, 5] <- 0; x[4, 6] <- 0; x[1, 5] <- 0
x
boot_table(x, n = 10)
boot_table(x, common_depth = 100)
boot_table(x, common_depth = 100, n = 10, round = TRUE)
boot_table(x, common_depth = 100, n = 10, replace_zero = 0.5)
}
