% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mvar_d3.R
\name{plot_mvar_d3}
\alias{plot_mvar_d3}
\title{Interactive multivariate analysis plots}
\usage{
plot_mvar_d3(mvar_object, types = NULL, height = NULL, width = NULL,
  rMin = NULL, rMax = NULL, ordinal_palettes = NULL,
  continuous_palettes = NULL, prop_input = NULL)
}
\examples{
library("FactoMineR")
data(wine)
wine_mfa <- ordi(wine[, -c(1:2)], "MFA", wine[, 1:2], group = c(5,3,10,9,2),
                 type = rep("s",5), graph = F)
plot_mvar_d3(wine_mfa)
plot_mvar_d3(wine_mfa, c("text", "arrow"), width = 500)
data(hobbies)
hobbies_mca <- ordi(hobbies[, 1:18], method = "MCA",
                    rows_annot = hobbies[, 19:23], graph = F)
plot_mvar_d3(hobbies_mca, width = 500, asp = 1)
}
