% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_mvar_d3.R
\name{plot_mvar_d3}
\alias{plot_mvar_d3}
\title{Interactive multivariate analysis plots}
\usage{
plot_mvar_d3(mvar_object, types = NULL, width = NULL, height = 400)
}
\description{
Interactive multivariate analysis plots
}

