% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Sample from Dirichlet Distribution}
\usage{
rdirichlet(alpha)
}
\arguments{
\item{alpha}{(Required). A vector of parameters for Dirichlet variable}
}
\value{
A vector of a random sample from Dirichlet distribution with
parameter \code{alpha}
}
\description{
Generate a sample from a dirichlet distribution with parameter \code{alpha}
}
\examples{
rdirichlet(sample(1:10, 5))
}

