% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vegan_to_mvar.R
\name{vegan_to_mvar}
\alias{vegan_to_mvar}
\title{Convert a vegan object into class mVarTable}
\usage{
vegan_to_mvar(vegan_object)
}
\arguments{
\item{An}{object resulting from a call to a \code{vegan} ordination method.}
}
\value{
mvar_table An object of class \code{mvarTable}, storing the site
   and species tables in the \code{table} slot and the eigenvalues in the
   \code{eig} slot. The annotation slots of each mVarLayer are the row names
   of the projected coordinates.
}
\description{
Convert a vegan object into class mVarTable
}

